+++ pallets/collator-selection/src/lib.rs
@@ -692,8 +692,6 @@ pub mod pallet {
 
     /// Fetch list of all possibly eligible authors to use in nimbus consensus filters
     ///
-    /// This is currently the static set registered with pallet_session and will be superseded by parachain_staking
-    ///
     /// NOTE: This should really be in pallet_session as we only use its storage, but since we haven't
     /// forked that one, this is the next best place.
     impl<T> Get<Vec<T::AccountId>> for Pallet<T>
