+++ runtime/calamari/Cargo.toml
@@ -93,6 +93,7 @@ manta-collator-selection = { path = '../../pallets/collator-selection', default-
 manta-primitives = { path = '../../primitives/manta', default-features = false }
 pallet-asset-manager = { path = '../../pallets/asset-manager', default-features = false }
 pallet-tx-pause = { path = '../../pallets/tx-pause', default-features = false }
+pallet-parachain-staking = { path = '../../pallets/parachain-staking', default-features = false }
 runtime-common = { path = '../common', default-features = false }
 session-key-primitives = { path = '../../primitives/session-keys', default-features = false }
 
@@ -132,6 +133,7 @@ runtime-benchmarks = [
   'calamari-vesting/runtime-benchmarks',
   'pallet-tx-pause/runtime-benchmarks',
   'pallet-treasury/runtime-benchmarks',
+  'pallet-parachain-staking/runtime-benchmarks',
   'pallet-preimage/runtime-benchmarks',
   'pallet-assets/runtime-benchmarks',
   'pallet-asset-manager/runtime-benchmarks',
@@ -145,6 +147,7 @@ try-runtime = [
   'pallet-aura/try-runtime',
   'pallet-authorship/try-runtime',
   'pallet-balances/try-runtime',
+  'pallet-parachain-staking/try-runtime',
   'pallet-preimage/try-runtime',
   'pallet-multisig/try-runtime',
   'pallet-session/try-runtime',
@@ -187,6 +190,7 @@ std = [
   'pallet-authorship/std',
   'pallet-balances/std',
   'pallet-multisig/std',
+  'pallet-parachain-staking/std',
   'pallet-preimage/std',
   'pallet-utility/std',
   'pallet-transaction-payment-rpc-runtime-api/std',
