+++ runtime/dolphin/src/lib.rs
@@ -24,16 +24,14 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
-use core::marker::PhantomData;
 use manta_collator_selection::IdentityCollator;
-use nimbus_primitives::AccountLookup;
 use sp_api::impl_runtime_apis;
 use sp_core::{crypto::KeyTypeId, OpaqueMetadata};
 use sp_runtime::{
     create_runtime_str, generic, impl_opaque_keys,
     traits::{AccountIdLookup, BlakeTwo256, Block as BlockT},
     transaction_validity::{TransactionSource, TransactionValidity},
-    ApplyExtrinsicResult, Perbill, Percent, Permill, RuntimeAppPublic,
+    ApplyExtrinsicResult, Perbill, Permill,
 };
 use sp_std::{cmp::Ordering, prelude::*};
 
@@ -41,12 +39,11 @@ use sp_std::{cmp::Ordering, prelude::*};
 use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 
-pub use frame_support::traits::Get;
 use frame_support::{
     construct_runtime, parameter_types,
     traits::{
-        ConstU128, ConstU16, ConstU32, ConstU8, Contains, Currency, EitherOfDiverse,
-        NeverEnsureOrigin, PrivilegeCmp,
+        ConstU16, ConstU32, ConstU8, Contains, Currency, EitherOfDiverse, NeverEnsureOrigin,
+        PrivilegeCmp,
     },
     weights::{
         constants::{BlockExecutionWeight, ExtrinsicBaseWeight, RocksDbWeight, WEIGHT_PER_SECOND},
@@ -62,7 +59,6 @@ use manta_primitives::{
     constants::{time::*, STAKING_PALLET_ID, TREASURY_PALLET_ID},
     types::{AccountId, Balance, BlockNumber, Hash, Header, Index, Signature},
 };
-pub use pallet_parachain_staking::{InflationInfo, Range};
 use runtime_common::{prod_or_fast, BlockHashCount, SlowAdjustingFeeUpdate};
 use session_key_primitives::{AuraId, NimbusId, VrfId};
 
@@ -280,60 +276,6 @@ impl Contains<Call> for BaseFilter {
     }
 }
 
-/// Converter struct to use the Session pallet for mapping a NimbusId to its AccountId
-pub struct TestConverter<T> {
-    _phantom: PhantomData<T>,
-}
-
-// Fetch list of eligible authors. This should be in manta_collator_selection
-impl<T> Get<Vec<T::AccountId>> for TestConverter<T>
-where
-    T: frame_system::Config + manta_collator_selection::Config + pallet_session::Config,
-    // Implemented only where Session's ValidatorId is directly convertible to collator_selection's ValidatorId
-    <T as manta_collator_selection::Config>::ValidatorId:
-        From<<T as pallet_session::Config>::ValidatorId>,
-{
-    /// Return the set of eligible collator accounts
-    fn get() -> Vec<T::AccountId>
-    where
-        <T as manta_collator_selection::Config>::ValidatorId:
-            From<<T as pallet_session::Config>::ValidatorId>,
-    {
-        use sp_runtime::traits::Convert;
-
-        let v = pallet_session::Pallet::<T>::validators()
-            .into_iter()
-            .map(|vid: <T as pallet_session::Config>::ValidatorId| {
-                <T as manta_collator_selection::Config>::AccountIdOf::convert(vid.into())
-            })
-            .collect::<Vec<T::AccountId>>();
-
-        log::info!("Requested Registered account Ids {:?}", v);
-        v
-    }
-}
-
-impl<T> AccountLookup<T::AccountId> for TestConverter<T>
-where
-    T: pallet_session::Config + manta_collator_selection::Config,
-    // Implemented only where Session's ValidatorId is directly convertible to collator_selection's ValidatorId
-    <T as manta_collator_selection::Config>::ValidatorId:
-        From<<T as pallet_session::Config>::ValidatorId>,
-{
-    fn lookup_account(author: &NimbusId) -> Option<T::AccountId>
-    where
-        <T as manta_collator_selection::Config>::ValidatorId:
-            From<<T as pallet_session::Config>::ValidatorId>,
-    {
-        use sp_runtime::traits::Convert;
-        #[allow(clippy::bind_instead_of_map)]
-        pallet_session::Pallet::<T>::key_owner(
-            nimbus_primitives::NIMBUS_KEY_ID,
-            &author.to_raw_vec(),
-        )
-        .and_then(|vid| Some(T::AccountIdOf::convert(vid.into())))
-    }
-}
 
 // Configure FRAME pallets to include in runtime.
 impl frame_system::Config for Runtime {
@@ -613,82 +555,9 @@ impl pallet_aura_style_filter::Config for Runtime {
     /// Nimbus filter pipeline (final) step 3:
     /// Choose 1 collator from PotentialAuthors as eligible
     /// for each slot in round-robin fashion
-    type PotentialAuthors = ParachainStaking;
+    type PotentialAuthors = CollatorSelection;
 }
 
-parameter_types! {
-    /// Default fixed percent a collator takes off the top of due rewards
-    pub const DefaultCollatorCommission: Perbill = Perbill::from_percent(20);
-    /// Default percent of inflation set aside for parachain bond every round
-    pub const DefaultParachainBondReservePercent: Percent = Percent::from_percent(30);
-}
-
-pub struct OnCollatorPayout;
-impl pallet_parachain_staking::OnCollatorPayout<AccountId, Balance> for OnCollatorPayout {
-    fn on_collator_payout(
-        _for_round: pallet_parachain_staking::RoundIndex,
-        _collator_id: AccountId,
-        _amount: Balance,
-    ) -> Weight {
-        // MoonbeamOrbiters::distribute_rewards(for_round, collator_id, amount)
-        0
-    }
-}
-pub struct OnNewRound;
-impl pallet_parachain_staking::OnNewRound for OnNewRound {
-    fn on_new_round(_round_index: pallet_parachain_staking::RoundIndex) -> Weight {
-        // MoonbeamOrbiters::on_new_round(round_index)
-        0
-    }
-}
-
-parameter_types! {
-    pub DefaultBlocksPerRound: BlockNumber = prod_or_fast!(2 * HOURS ,15,"DOLPHIN_DEFAULTBLOCKSPERROUND");
-}
-impl pallet_parachain_staking::Config for Runtime {
-    type Event = Event;
-    type Currency = Balances;
-    type BlockAuthor = AuthorInherent;
-    type MonetaryGovernanceOrigin = EnsureRoot<AccountId>;
-    /// Minimum round length is 2 minutes (10 * 12 second block times)
-    type MinBlocksPerRound = ConstU32<10>;
-    /// Blocks per round
-    type DefaultBlocksPerRound = DefaultBlocksPerRound;
-    /// Rounds before the collator leaving the candidates request can be executed
-    type LeaveCandidatesDelay = ConstU32<24>;
-    /// Rounds before the candidate bond increase/decrease can be executed
-    type CandidateBondLessDelay = ConstU32<24>;
-    /// Rounds before the delegator exit can be executed
-    type LeaveDelegatorsDelay = ConstU32<24>;
-    /// Rounds before the delegator revocation can be executed
-    type RevokeDelegationDelay = ConstU32<24>;
-    /// Rounds before the delegator bond increase/decrease can be executed
-    type DelegationBondLessDelay = ConstU32<24>;
-    /// Rounds before the reward is paid
-    type RewardPaymentDelay = ConstU32<2>;
-    /// Minimum collators selected per round, default at genesis and minimum forever after
-    type MinSelectedCandidates = ConstU32<8>;
-    /// Maximum top delegations per candidate
-    type MaxTopDelegationsPerCandidate = ConstU32<300>;
-    /// Maximum bottom delegations per candidate
-    type MaxBottomDelegationsPerCandidate = ConstU32<50>;
-    /// Maximum delegations per delegator
-    type MaxDelegationsPerDelegator = ConstU32<100>;
-    type DefaultCollatorCommission = DefaultCollatorCommission;
-    type DefaultParachainBondReservePercent = DefaultParachainBondReservePercent;
-    /// Minimum stake required to become a collator
-    type MinCollatorStk = ConstU128<{ 1000 * DOL }>;
-    /// Minimum stake required to be reserved to be a candidate
-    type MinCandidateStk = ConstU128<{ 500 * DOL }>;
-    /// Minimum stake required to be reserved to be a delegator
-    type MinDelegation = ConstU128<{ 5 * DOL }>;
-    /// Minimum stake required to be reserved to be a delegator
-    type MinDelegatorStk = ConstU128<{ 5 * DOL }>;
-    type OnCollatorPayout = OnCollatorPayout;
-    type OnNewRound = OnNewRound;
-    type WeightInfo = ();
-    // type WeightInfo = pallet_parachain_staking::weights::SubstrateWeight<Runtime>; TODO
-}
 
 impl pallet_author_inherent::Config for Runtime {
     // We start a new slot each time we see a new relay block.
@@ -847,7 +716,6 @@ construct_runtime!(
         TechnicalCommittee: pallet_collective::<Instance2>::{Pallet, Call, Storage, Origin<T>, Event<T>, Config<T>} = 17,
         TechnicalMembership: pallet_membership::<Instance2>::{Pallet, Call, Storage, Event<T>, Config<T>} = 18,
 
-        ParachainStaking: pallet_parachain_staking::{Pallet, Call, Storage, Event<T>, Config<T>} = 48,
         // Collator support.
         AuthorInherent: pallet_author_inherent::{Pallet, Call, Storage, Inherent} = 60,
         AuraAuthorFilter: pallet_aura_style_filter::{Pallet, Storage} = 63,
@@ -955,7 +823,6 @@ mod benches {
         [manta_collator_selection, CollatorSelection]
         [pallet_manta_pay, MantaPay]
         [pallet_asset_manager, AssetManager]
-        [pallet_parachain_staking, ParachainStaking]
         // Nimbus pallets
         [pallet_author_inherent, AuthorInherent]
 
@@ -969,7 +836,9 @@ impl_runtime_apis! {
         }
 
         fn authorities() -> Vec<AuraId> {
-            Aura::authorities().into_inner()
+            // NOTE: AuraAPI must exist for node/src/aura_or_nimbus_consensus.rs
+            // But is intentionally DISABLED starting with manta v3.3.0
+            vec![]
         }
     }
 
@@ -1188,12 +1057,4 @@ cumulus_pallet_parachain_system::register_validate_block! {
     Runtime = Runtime,
     BlockExecutor = pallet_author_inherent::BlockExecutor::<Runtime, Executive>,
     CheckInherents = CheckInherents,
-}
-
-// Shorthand for a Get field of a pallet Config.
-#[macro_export]
-macro_rules! get {
-    ($pallet:ident, $name:ident, $type:ty) => {
-        <<$crate::Runtime as $pallet::Config>::$name as $crate::Get<$type>>::get()
-    };
 }
